refactor(将resourceLoader相关函数移动至单独的文件):
