+++ components/ExternalPlugins.js
@@ -1,6 +1,5 @@
 import { siteConfig } from '@/lib/config'
 import { convertInnerUrl } from '@/lib/notion/convertInnerUrl'
-import { loadExternalResource } from '@/lib/utils'
 import dynamic from 'next/dynamic'
 import { useRouter } from 'next/router'
 import { useEffect } from 'react'
@@ -13,6 +12,7 @@ import WebWhiz from './Webwhiz'
 import { useGlobal } from '@/lib/global'
 import IconFont from './IconFont'
 import { isBrowser } from '@/lib/utils/envUtils'
+import { loadExternalResource } from '@/lib/utils/resourceLoader'
 
 
 /**
