+++ components/Fireworks.js
@@ -5,7 +5,8 @@
 import { useEffect } from 'react'
 // import anime from 'animejs'
 import { siteConfig } from '@/lib/config'
-import { loadExternalResource } from '@/lib/utils'
+
+import { loadExternalResource } from '@/lib/utils/resourceLoader'
 
 /**
  * 鼠标点击烟花特效
