+++ components/Live2D.js
@@ -1,9 +1,9 @@
 /* eslint-disable no-undef */
 import { siteConfig } from '@/lib/config'
 import { useGlobal } from '@/lib/global'
-import { loadExternalResource } from '@/lib/utils'
 import { useEffect } from 'react'
 import { isMobile } from '@/lib/utils/envUtils'
+import { loadExternalResource } from '@/lib/utils/resourceLoader'
 
 /**
  * 网页动画
