+++ components/NotionPage.js
@@ -1,12 +1,12 @@
 import { siteConfig } from '@/lib/config'
 import { compressImage, mapImgUrl } from '@/lib/notion/mapImage'
-import { loadExternalResource } from '@/lib/utils'
 import mediumZoom from '@fisch0920/medium-zoom'
 import 'katex/dist/katex.min.css'
 import dynamic from 'next/dynamic'
 import { useEffect, useRef } from 'react'
 import { NotionRenderer } from 'react-notion-x'
 import { isBrowser } from '@/lib/utils/envUtils'
+import { loadExternalResource } from '@/lib/utils/resourceLoader'
 
 /**
  * 整个站点的核心组件
