+++ components/Ribbon.js
@@ -1,5 +1,6 @@
 import { useEffect } from 'react'
-import { loadExternalResource } from '@/lib/utils'
+
+import { loadExternalResource } from '@/lib/utils/resourceLoader'
 
 const Ribbon = () => {
   useEffect(() => {
