+++ components/ValineComponent.js
@@ -1,6 +1,6 @@
 import { siteConfig } from '@/lib/config'
-import { loadExternalResource } from '@/lib/utils'
 import { useEffect } from 'react'
+import { loadExternalResource } from '@/lib/utils/resourceLoader'
 
 const ValineComponent = ({ path }) => {
   const loadValine = async () => {
