+++ lib/utils/resourceLoader.js
@@ -1,0 +1,47 @@
+/**
+ * 加载外部资源
+ * @param url 地址 例如 https://xx.com/xx.js
+ * @param type js 或 css
+ * @returns {Promise<unknown>}
+ */
+export function loadExternalResource(url, type = 'js') {
+  // 检查是否已存在
+  const elements =
+    type === 'js'
+      ? document.querySelectorAll(`[src='${url}']`)
+      : document.querySelectorAll(`[href='${url}']`)
+
+  return new Promise((resolve, reject) => {
+    if (elements.length > 0 || !url) {
+      resolve(url)
+      return url
+    }
+
+    let tag
+
+    if (type === 'css') {
+      tag = document.createElement('link')
+      tag.rel = 'stylesheet'
+      tag.href = url
+    } else if (type === 'font') {
+      tag = document.createElement('link')
+      tag.rel = 'preload'
+      tag.as = 'font'
+      tag.href = url
+    } else if (type === 'js') {
+      tag = document.createElement('script')
+      tag.src = url
+    }
+    if (tag) {
+      tag.onload = () => {
+        // console.log('Load Success', url)
+        resolve(url)
+      }
+      tag.onerror = () => {
+        console.warn('Load Error', url)
+        reject(url)
+      }
+      document.head.appendChild(tag)
+    }
+  })
+}
