+++ themes/hexo/components/Hero.js
@@ -2,10 +2,10 @@
 import LazyImage from '@/components/LazyImage'
 import { siteConfig } from '@/lib/config'
 import { useGlobal } from '@/lib/global'
-import { loadExternalResource } from '@/lib/utils'
 import { useEffect, useState } from 'react'
 import CONFIG from '../config'
 import NavButtonGroup from './NavButtonGroup'
+import { loadExternalResource } from '@/lib/utils/resourceLoader'
 
 let wrapperTop = 0
