+++ themes/matery/components/Hero.js
@@ -2,9 +2,9 @@
 import LazyImage from '@/components/LazyImage'
 import { siteConfig } from '@/lib/config'
 import { useGlobal } from '@/lib/global'
-import { loadExternalResource } from '@/lib/utils'
 import { useEffect, useState } from 'react'
 import CONFIG from '../config'
+import { loadExternalResource } from '@/lib/utils/resourceLoader'
 
 let wrapperTop = 0
