+++ themes/next/components/Live2DWaifu.js
@@ -1,6 +1,7 @@
 import Head from 'next/head'
 import { useEffect } from 'react'
-import { loadExternalResource } from '@/lib/utils'
+
+import { loadExternalResource } from '@/lib/utils/resourceLoader'
 
 export default function Live2DWife() {
   useEffect(() => {
